+++ src/Utils/ExternalEmittance.cpp
