+++ src/Utils/ExternalEmittance.h
@@ -1,5 +1,10 @@
 #pragma once
 
+namespace RE
+{
+	class BSRenderPass;
+}
+
 namespace ExternalEmittance
 {
 	bool ShouldSuppress(const RE::BSRenderPass* a_pass);
